reverted Java 9+ privateLookup
